Fixes bug #113 where Params are incorrectly converted to hashes.
The problem occurs when we're using whitelisting and nested
parameters.  When attribute definitions are created in the context
of Lotus::Action::Params the anonymous Params subclass which is
created does not respond to the :lotus_nested_attributes? method
which is expected by Lotus::Utils::Attribues#to_h.

The fix is to make the anonymous class respond to that method with
the right value.  This is a little hacky for two reasons:

1. You need to look at the source for
Lotus::Validations::AttributeDefiner to understand this.
2. It might make sense to create a NestParams class in
Lotus::Action.