Rework #accepts to be more clear
This is meant to clarify the intent and overall create a more focused and
useful method. Realistically, this is a private API for the class, and it's
only used in one place - but nonetheless the intent and purpose should be
clear.

This is now (clearly) responsible for looking at the current Accept header
value (default or not) and identifying (with weight compensation) if the
value is a known (non-custom) MIME type. If it is, the type is returned,
otherwise nil is returned.