Add #accept_header?
This method is simply replacing the logic that used to exist in the methods
that look to the header for a certain MIME type (either common or custom).
Now, instead of checking for the header to exist, then looking at type, they
are simpler and only look at type.

The check for the Accept headers presence is now taken care of external to
those methods (since they shouldn't be aware of that state anyway).