Honor RFC 2616 (14.1) with Accept header priority
In my earlier commits the preference for Content-Type resolution was focused
on the types that are common (from Rack), THEN on the types that were
registered through the controller configuration. Now we are going to look at
both of those with the `q` weighting that is part of the RFC (and currently
part of how Hanami Controller works).