Add accessor to configuration for known MIME types
We need to be able to know what MIME types have been registered with a custom
controller configuration, so we can appropriately get the correct Content-Type
to respond with (given an Accept header).

There was no accessor method other than the protected `#formats` which only
returns a hash including both MIME type and format key (we only need the
former).