Don't prepend `#call`, instead require subclasses to implement `#handle`
This improves understandability of the library - when a subclass implements `#handle`, but the main interface remains `#call`, then application developers will implicitly understand that some extra code is being run in between the two.

It also removes the need for the module prepend to be run in the `.inherited` hook, which causes problem for deeper inheritance hierarchies, since it would prepend the module multiple times.