Remove class-level handle_exception method
This is already covered by the class-level config object now (i.e. config.handle_exception or configuration.handle_exception, if you prefer a bit more typing)

Further, unlike the config, the class-level handled exceptions were not inherited by subclasses.

This is a hangover from the previous design that didn’t get tidied up earlier.

Along with this, update all specs to use class-level config rather than injecting it explicitly (which is not the primary way to use it).