Enable CSRF protection automatically (#332)
Enable CSRF protection automatically on application actions if the application config has `sessions.enabled?`.

To support this:

- Move the `sessions` config into hanami-controller, so it is accessible via `config.actions.sessions` (as part of `ApplicationConfiguration`)
- Add a `csrf_protection` setting to `ApplicationConfiguration`, defaulting to `nil`
- Use `ApplicationConfiguration#finalize!` to set `csrf_protection` to true or false based on whether sessions are enabled (but _only_ if it was previously `nil`, which allows the application author to explicitly disable it if they desire)

This relies upon https://github.com/hanami/hanami/pull/1078, which calls `finalize!` on the `config.actions` and `config.views` objects.

In addition to this, it make it possible to use _all_ the public methods on methods on `Hanami::Action::Configuration` when accessing it via an `ApplicationConfiguration`, i.e. via `Hanami.application.config.actions`. To support this, we now delegate to `Hanami::Action::Configuration` as a "base configuration" from within `Hanami::Action::ApplicationConfiguration`, which means all methods work now, not just the defined settings.