Late bind default value of root_directory
Previously, putting the default of Dir.pwd directly into the definition of the `root_directory` setting meant that in the case of the Hanami framework integration tests, where applications are created, torn down, and recreated in new working directories, a stale root_directory would be used, and result in errors since its no longer an existent path.

Adjusting this setting to default to Dir.pwd _inside_ the setting’s processor still results in the same default being applied, just without the downsides of it being captured early and persisted.