Fix mutable state.
By using `default:` the `ContentSecurityPolicy` was cached at the class level at the load time.
All the new instances of `ApplicationConfiguration` was always using the same instance of `ContentSecurityPolicy`.

If one of the tests was altering the CSP value of a configuration
instance, that change was reflected in other configuration instances as
well, because they were sharing the same `ContentSecurityPolicy`
instance.

`constructor:` accepts a `Proc`, which is evaluated at the init time of
`ApplicationConfiguration`. That results into the expected behavior:
each configuration instance has its own instance of
`ContentSecurityPolicy`.