Change ApplicationAction to a class
Before this, inheriting from `Hanami::Action` would automagically
integrate with Hanami::Application if it was done inside an app.

With this change, users will instead inherit from `Hanami::ApplicationAction`

While this is slightly less convenient, Hanami 2 leverages inheritance for
project and slice-specific superclasses, so changing it in one place
seems like a worthwhile trade-off for library code (this file) being
easier to understsand (for users diving in) and maintain (for contributors).

We're also able to leverage good OO design, class inheritance, and
dependency injection with keyword args.