fix: relaxed mime types are the second argument.
From a look at the source of Rack::Mime.match?, wildcard arguments (e.g.
*/* or text/*) will only be treated as actual wildcards when given as
the second argument.

So if in a Hanami action we state we want to accept all content via
`format :all`, the corresponding default mapping of */* needs to be
supplied as the second argument here to ensure requests are actually
accepted.

https://github.com/rack/rack/blob/v3.0.3/lib/rack/mime.rb#L23-L35