Add the action instance to the request environment
I created a [new topic] on the Hanami Discourse about improving the
AppSignal APM instrumentation.

To group requests, we would like to know what action it took place in.
This information is currently not available in the request environment.

For us, the easiest would be to access the action instance. That gives
us the class information and a way to access the `params_class` to fetch
all the parameters of the request: query params and the body payload.

This change adds the action instance on a new request environment key
`hanami.action_instance`.

[new topic]: https://discourse.hanamirb.org/t/questions-for-improving-the-appsignal-apm-integration-with-hanami-2/989/3

I had to update one test to not fail on the action instance being
returned, which is not important for that spec I think.

Closes #445