Make config.formats clearer and more flexible (#485)
Provide new methods on the `config.formats` object that make it more flexible and easier to work with.

These are _all new methods_, intended to become the recommended usage of `config.formats` from Hanami 2.3 onwards.

For the purposes of compatibility, the existing methods remain in place, but they now emit deprecation warnings and are no longer documented. We will include notes on how to switch to the new methods in the upgrade guide.

### Add `formats.register`, with types for distinct purposes

`formats.register` replaces `formats.add`. Its purpose is to register format to media type mappings _only_. Unlike `.add`, it does _not_ set any of the given formats as being “accepted” by the action (see next section for the new way to do this).

This change makes it easier to `register` a range of formats e.g. in a base action class or in the app-level `config.actions.formats`, while ensuring you do not cause any inadvertent format restrictions in child action classes.

Example:

```ruby
config.formats.register(:json, "application/json")
```

`.register` also allows you to register one or more media types for the distinct stages of request processing.

- If you want to accept requests based on different/additional media types in `Accept` request headers, provide them as `accept_types:`
- If you want to accept requests based on different/additional media types in `Content-Type` request headers, provide them as `content_types:`
- If you do not provide these options, then the _default_ media type (the required second argument, after the format name) is used for each of the above
- This default media type is also set as the default `Content-Type` _response_ header for requests that match the format

Example:

```ruby
config.formats.register(
  :jsonapi,
  "application/vnd.api+json",
  accept_types: ["application/vnd.api+json", "application/json"],
  content_types: ["application/vnd.api+json", "application/json"],
)
```

### Add `formats.accepted` and `formats.accept`

`formats.accepted` replaces `formats.values`. This name better reflects its purpose: it is the array of formats that the action will _accept_ when handling requests.

To populate this array, you can still manually assign it, e.g. `formats.accepted = [:json]` as you could do previously with `#values=`.

However, the nicer (and recommended) way to do this is via `config.accept`, which takes a splat of format names. Example:

```ruby
config.formats.accept :html, :json
config.formats.accepted # => [:html, :json]

config.formats.accept :csv # it is additive
config.formats.accepted # => [:html, :json, :csv]
```

This method is also a replacement for the higher-level `config.format` method that we currently have documented. In making these changes, I decided that it's better to teach users about the existence of the `config.formats` object and its own methods rather than trying to hide it away behind a single (and therefore inflexible) convenience method.

### Add `formats.default=`

This is a brand new capability: you can now assign a _default_ format via e.g. `formats.default = :json`.

The default format is used to set the default `Content-Type` response header if one of the configured accepted formats _cannot_ be matched to the request's `Accept` header.

Previously, the default format was set to be the first of the formats that the `config.formats` becomes aware of, either via running `config.formats.add` or `config.format` or `config.formats.values=`. This was a bad arrangement because the setting of the default was _implicit_. It was not obvious from any of the methods being called and led to unexpected surprises. It could also not easily be overridden, which made things awkward in situations where action config is inherited from parent classes or app-level config.

Now we have a simple setter that allows the default format to be configured independently.

I should note that I have preserved the automatic setting of a default when calling `formats.accept`, but _only_ if the `formats.default` is not already set with a non-nil value. This way we hopefully have the best of both worlds: a one-liner when someone wants it, but the ability to explicitly and separately configure the default format when that is required.

### Use “media type” naming

I've also changed all references to "MIME types" to "media types", since that is the modern term for these. The only places I've left "MIME" in place is for our own MIME module (which is private API anyway) and any MIME-related methods we call or import from Rack.